+++ tests/fixtures/network.py
@@ -3,7 +3,7 @@ import docker.errors
 import pytest
 from testcontainers.core.network import Network
 
-from fixtures import dev, types
+from fixtures import reuse, types
 from fixtures.logger import setup_logger
 
 logger = setup_logger(__name__)
@@ -37,7 +37,7 @@ def network(
         nw = client.networks.get(network_id=existing.get("id"))
         return types.Network(id=nw.id, name=nw.name)
 
-    return dev.wrap(
+    return reuse.wrap(
         request,
         pytestconfig,
         "network",
