+++ tests/fixtures/zookeeper.py
@@ -3,7 +3,7 @@ import docker.errors
 import pytest
 from testcontainers.core.container import DockerContainer, Network
 
-from fixtures import dev, types
+from fixtures import reuse, types
 from fixtures.logger import setup_logger
 
 logger = setup_logger(__name__)
@@ -58,7 +58,7 @@ def zookeeper(
     def restore(cache: dict) -> types.TestContainerDocker:
         return types.TestContainerDocker.from_cache(cache)
 
-    return dev.wrap(
+    return reuse.wrap(
         request,
         pytestconfig,
         "zookeeper",
